+++ NBE_5_7_2_02TEAM/src/main/java/io/twogether/nbe_5_7_2_02team/chat/api/ChatMemberController.java
@@ -1,7 +1,7 @@
 package io.twogether.nbe_5_7_2_02team.chat.api;
 
-import io.twogether.nbe_5_7_2_02team.chat.dto.ChatMemberGetResponse;
-import io.twogether.nbe_5_7_2_02team.chat.dto.ChatMemberUpdateRequest;
+import io.twogether.nbe_5_7_2_02team.chat.dto.response.ChatMemberGetResponse;
+import io.twogether.nbe_5_7_2_02team.chat.dto.request.ChatMemberUpdateRequest;
 import io.twogether.nbe_5_7_2_02team.chat.service.ChatMemberService;
 import io.twogether.nbe_5_7_2_02team.global.response.success.BaseResponse;
 import io.twogether.nbe_5_7_2_02team.global.response.success.SuccessCode;
