+++ NBE_5_7_2_02TEAM/src/main/java/io/twogether/nbe_5_7_2_02team/chat/dto/request/ChatMessagePostRequest.java
@@ -1,4 +1,4 @@
-package io.twogether.nbe_5_7_2_02team.chat.dto;
+package io.twogether.nbe_5_7_2_02team.chat.dto.request;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
