+++ NBE_5_7_2_02TEAM/src/test/java/io/twogether/nbe_5_7_2_02team/chat/service/ChatMessageServiceTest.java
@@ -13,8 +13,8 @@ import io.twogether.nbe_5_7_2_02team.chat.dao.ChatMemberRepository;
 import io.twogether.nbe_5_7_2_02team.chat.dao.ChatMessageRepository;
 import io.twogether.nbe_5_7_2_02team.chat.dao.ChatRoomRepository;
 import io.twogether.nbe_5_7_2_02team.chat.domain.ChatRoom;
-import io.twogether.nbe_5_7_2_02team.chat.dto.ChatMessageGetResponse;
-import io.twogether.nbe_5_7_2_02team.chat.dto.ChatMessagePostRequest;
+import io.twogether.nbe_5_7_2_02team.chat.dto.response.ChatMessageGetResponse;
+import io.twogether.nbe_5_7_2_02team.chat.dto.request.ChatMessagePostRequest;
 import io.twogether.nbe_5_7_2_02team.global.exception.ErrorException;
 import io.twogether.nbe_5_7_2_02team.member.dao.MemberRepository;
 import io.twogether.nbe_5_7_2_02team.member.domain.Member;
